Refactor Tone component
